pyproject.toml
[tool.poetry]
name = "anns-war"
version = "0.1.0"
description = ""
authors = ["Nawat Ngerncham <n.ngerncham@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "1.25.0"
hnswlib = "0.7.0"
faiss-cpu = "1.7.4"
diskannpy = "^0.6.1"
psutil = "^5.9.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
